[app] Understanding the task
It turns out that i have to make a managment system, not a delivery
system. The good part is, i did this early enought to fix all this.

I'm adding admin account, and the permission will be tied to it's name,
because the app doesn't seem complex enough to be role-based. They
simply need a way to keep track of what they have.

On another note, i should make my home page an instructions page, since
this website is used only within the bounds of the company.